rust-server-cleanup(utils,non_mechanical_provable): extract boundary helpers
