+++ rust/sglang-server/src/utils.rs
@@ -8,5 +8,6 @@ pub mod regex;
 pub mod response;
 pub mod runtime;
 pub mod serialize;
+pub mod server;
 pub mod sock;
 pub mod threads;
